requirements.txt
streamlit>=1.28.0
agno>=0.2.0
google-generativeai>=0.3.2
apify-client>=1.6.0
pydantic>=2.0.0
python-dotenv>=1.0.0
pandas>=2.0.0
numpy>=1.24.0
requests>=2.31.0
beautifulsoup4>=4.12.0
tiktoken>=0.5.0 